refactor: Move logging.py to main module to avoid circular import issues
