+++ archivebox/cli/archivebox_config.py
@@ -10,7 +10,7 @@ from typing import Optional, List, IO
 
 from ..main import config, docstring
 from ..config import OUTPUT_DIR
-from .logging import SmartFormatter, accept_stdin
+from ..logging import SmartFormatter, accept_stdin
 
 
 @docstring(config.__doc__)
