+++ archivebox/cli/archivebox_server.py
@@ -10,7 +10,7 @@ from typing import Optional, List, IO
 
 from ..main import server, docstring
 from ..config import OUTPUT_DIR
-from .logging import SmartFormatter, reject_stdin
+from ..logging import SmartFormatter, reject_stdin
 
 
 @docstring(server.__doc__)
